@@ -1,8 +1,10 @@
 import React from "react";
 import './index.css';
+import TodoTemplate from './components/TodoTemplate';
+import TodoInsert from './components/TodoInsert';
 
 const App = () => {
-	return <div>Todo App을 만들자</div>;
+	return <TodoTemplate><TodoInsert /></TodoTemplate>;
 };
 
 export default App;
